+++ tests/src/Calendar/CalendarTest.php
@@ -4,15 +4,15 @@ namespace garethp\ews\Test\Calendar;
 
 use garethp\ews\Test\BaseTestCase;
 
-class APITest extends BaseTestCase
+class CalendarTest extends BaseTestCase
 {
-    public function setUp()
+    public function setUp(): void
     {
         $client = $this->getClient();
         $client->deleteAllCalendarItems('2015-07-01 00:00', '2015-07-01 23:59');
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         $client = $this->getClient();
         $client->deleteAllCalendarItems('2015-07-01 00:00', '2015-07-01 23:59');